added multiple candidates to fineYIN pass

+++ include/audioprocessing.h
@@ -3,12 +3,12 @@
 
 #include "config.h"
 
-// Audio processing function declarations
+// audio subsystem and analysis functions
 bool initI2S();
 bool captureRealAudio(AudioBuffer* buffer);
 int findCoarsePeriodYIN(const AudioBuffer* input, TuningResult* result);
 bool yinAnalysis(const AudioBuffer* input, TuningResult* output, int hintedPeriod);
 void convertFrequencyToNote(float frequency, char* noteName, size_t nameSize);
-float calculateCentsOffset(float frequency);
+int calculateCentsOffset(float frequency);
 
 #endif // AUDIOPROCESSING_H
